requirements.txt
pytest
pytest-mock
python-statemachine
python-open-weather
Telethon==1.27.0
gspread
python-dateutil
retry==0.9.2
python-telegram-bot==13.14
matplotlib==3.7.2